Fix PPS by retaining the File
I was keeping the File in PPS object, then I extracted its fd and passed
the fd into a future.  The PPS object was dropped so the fd became
invalid leading to EBADF when attempting to read a PPS event.

Now (I think) the file is held in the future and will keep its fd alive.